requirements.txt
Flask>=0.10.1
peewee>=2.1.2
flask-peewee>=0.6.5
python-ldap>=2.4.12
MySQL-python>=1.2.4
lxml>=3.2.1
html5lib>=1.0b1
beautifulsoup4>=4.2.1
wtf-peewee>=0.2.2
WTForms>=1.0.4
Jinja2>=2.7
Werkzeug>=0.9.1
repoze.profile>=2.0
